JuegoPantalla con player y movimiento
se agrego todo lo necesario para juego pantalla, se dibujo al player con su movimiento basico

+++ core/src/main/java/com/cartoonsurvivors/game/entidades/Jugador.java
@@ -1,7 +1,7 @@
 package com.cartoonsurvivors.game.entidades;
 
 public class Jugador extends Entidad {
-    public Jugador(int vida, double posicionX, double posicionY) {
+    public Jugador(int vida, int posicionX, int posicionY) {
         super(vida, posicionX, posicionY);
     }
